Add a flags field to the pdu structure Add a new pdu flag : DELETE_WHEN_SENT. When this pdu has been sent to the wire, the pdu will be deleted and not put on the waitpdu list.
This will be useful for sequences such as
    -> WRITE10 cdb
    -> DATAOUT the data to write
    -< RESPONSE

Where we want to match WRITE10 and RESPONSE but where the plain DATAOUT pdu
will not be soliciting its own response.
We dont need to wait for the response to DATAOUT pdus, we are already waitin
for the response form the initial PDU in the sequence.