socket: add support for slave contexts using epoll
Slave contexts are attached to a master context. All slave contexts are
serviced from the master context they are attached to.
This makes managing of multiple contexts more simple since instead of having
to manage iscsi_get_fd()/iscsi_which_events()/iscsi_service() for each
and every context, you can attach all but one context as a slave to the master
context and then only have to manage the file descriptor for the master context.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>