Allow sending TUR to consume all UAs during reconnect
During reconnect we normally defer any SCSI commands that are issued
to be queued and sent later, once the re-connect has completed and we have
swapped the contexts.
This is what we want for almost all situations, except when we are
reconnecting very simple applications which request "no ua on reconnect".
For these applications we want to actually send the TURs that are used
during the login phase on the temporary context.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>